+++ lib/pages/settings_page.dart
@@ -18,6 +18,7 @@ import 'package:remote_storage/widgets/settings_sections.dart'
         VisibilitySection,
         WebDavCredentialsSection;
 import 'package:remote_storage/widgets/settings_cache_section.dart';
+import 'package:remote_storage/widgets/settings_log_section.dart';
 import 'package:remote_storage/widgets/settings_proxy_section.dart';
 import 'package:remote_storage/widgets/settings_sync_section.dart';
 import 'package:remote_storage/widgets/settings_trash_section.dart';
@@ -36,6 +37,7 @@ enum _SettingsTab {
   update,
   proxy,
   appearance,
+  logging,
   download,
   cache,
   visibility,